LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/evaluation-error-4-expected.txt

CONSOLE MESSAGE: [object Object]

FAIL Test that exceptions during evaluation lead to error events on window, and that exceptions are remembered.
 assert_array_equals: lengths differ, expected array ["throw", object "[object Object]", [...], [...], [...], [...]] length 6, got ["throw", object "[object Object]"] length 2
